school/seriallizers/payments.py
from rest_framework import serializers

from school.models import Payments


class PaymentsSerializer(serializers.ModelSerializer):
    class Meta:
        model = Payments
        fields = '__all__'
